Upgrade global exception handler to utilize .NET 8's IExceptionHandler
