--- application/shared-kernel/ApiCore/Middleware/GlobalExceptionHandlerMiddleware.cs
@@ -1,43 +1,0 @@
-using System.Net;
-using System.Text.Json;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Options;
-using JsonOptions = Microsoft.AspNetCore.Http.Json.JsonOptions;
-
-namespace PlatformPlatform.SharedKernel.ApiCore.Middleware;
-
-public sealed class GlobalExceptionHandlerMiddleware(
-    ILogger<GlobalExceptionHandlerMiddleware> logger,
-    IOptions<JsonOptions> jsonOptions
-) : IMiddleware
-{
-    private readonly JsonSerializerOptions _jsonSerializerOptions = jsonOptions.Value.SerializerOptions;
-    private readonly ILogger _logger = logger;
-
-    public async Task InvokeAsync(HttpContext context, RequestDelegate next)
-    {
-        try
-        {
-            await next(context);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "An error occurred while processing the request.");
-
-            context.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
-            context.Response.ContentType = "application/problem+json";
-
-            var problemDetails = new ProblemDetails
-            {
-                Type = $"https://httpstatuses.com/{(int)HttpStatusCode.InternalServerError}",
-                Title = "Internal Server Error",
-                Status = (int)HttpStatusCode.InternalServerError,
-                Detail = "An error occurred while processing the request."
-            };
-
-            var jsonResponse = JsonSerializer.Serialize(problemDetails, _jsonSerializerOptions);
-            await context.Response.WriteAsync(jsonResponse);
-        }
-    }
-}
